refactor(ecommerce): abstract header into it's own component
